requirements.txt
azure-ai-vision-imageanalysis>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
